autogen-mcp-agent-test/requirements.txt
autogen
autogen-mcp>=0.1.0
python-dotenv>=1.0.0
